requirements.txt
asgiref==3.5.2
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Django==4.0.5
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
executing==0.8.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.12.0
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
traitlets==5.2.2.post1
tzdata==2022.1
wcwidth==0.2.5
